Results and Meaning component
added new use state to dictionary component for results , sending them as props to a new component called Results. Results displays the word that was searched for and maps through word definitions. creating a new div for each, sending to Meaning component. the Meaning component receives the definitions one by one and displays.